PoA Improvements #2 (#8991)
* Code improvements and remove shuffleValidatorList from PoA

* Revert changes that has already been resolved in #8967 at update_generator_key

---------
