+++ framework/src/modules/pos/module.ts
@@ -70,12 +70,11 @@ import {
 	equalUnlocking,
 	isUsername,
 	selectStandbyValidators,
-	shuffleValidatorList,
 	sortUnlocking,
 	getModuleConfig,
 	getValidatorWeight,
-	ValidatorWeight,
 	isSharingCoefficientSorted,
+	ValidatorWeight,
 } from './utils';
 import { ValidatorStore } from './stores/validator';
 import { GenesisDataStore } from './stores/genesis';
@@ -93,6 +92,7 @@ import { CommissionChangeEvent } from './events/commission_change';
 import { ClaimRewardsCommand } from './commands/claim_rewards';
 import { getMainchainID } from '../interoperability/utils';
 import { RewardsAssignedEvent } from './events/rewards_assigned';
+import { shuffleValidatorList } from '../utils';
 
 export class PoSModule extends BaseModule {
 	public method = new PoSMethod(this.stores, this.events);
@@ -690,7 +690,7 @@ export class PoSModule extends BaseModule {
 		}
 
 		// Update the validators
-		const shuffledValidators = shuffleValidatorList(randomSeed1, validators);
+		const shuffledValidators = shuffleValidatorList<ValidatorWeight>(randomSeed1, validators);
 		let aggregateBFTWeight = BigInt(0);
 		const bftValidators: { address: Buffer; bftWeight: bigint }[] = [];
 		for (const v of shuffledValidators) {
